xlm/requirements-adapter.txt
# Automatically generated by https://github.com/damnever/pigar.

adapter-transformers==3.2.1
datasets==2.14.5
entmax==1.1
evaluate==0.4.0
fastdist==1.1.6
fasttext==0.9.2
lightning==2.0.9.post0
loguru==0.7.2
numpy==1.24.4
packaging==23.2
pandas==2.0.3
requests==2.31.0
sentencepiece==0.1.99
seqeval==1.2.2
tokenizers==0.13.3
tqdm==4.66.1
wandb==0.15.12

--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0